Req OLTSWITCH-1341 Support the card TC16SXG - new toolchain (phase 3/4):  - Corrected errors regarding to the usage of size_t variables;  - Corrected callback functions wich should use size_t parameters instead of L7_uint32.
svn path=/trunk/; revision=7748

+++ src/fastpath/os/linux/mgmt/dtl_tap_monitor.c
@@ -396,7 +396,7 @@ void tap_monitor_task_fn()
              *file descriptor
              */
             frame_length = read(reg_table[i].fd,(void *)frame_data,MAX_FRAME_LEN);
-            PRINT_CTRL_MSG("frame length is %d bytes\n",frame_length);
+            PRINT_CTRL_MSG("frame length is %u bytes\n", (unsigned int) frame_length);
 
             /*
              *check to make sure we got something
